Refactor part 3
