+++ .babelrc
@@ -1,0 +1,24 @@
+{
+    "ignore": [
+        "node_modules/**/*"
+    ],
+    "presets": [
+        [
+            "@babel/env",
+            {
+                "modules": "umd"
+            }
+        ]
+    ],
+    "plugins": [
+        "@babel/plugin-proposal-class-properties",
+        [
+            "@babel/plugin-transform-modules-umd",
+            {
+                "globals": {
+                    "jquery": "jQuery"
+                }
+            }
+        ]
+    ]
+}
