+++ package.json
@@ -20,6 +20,8 @@
   "homepage": "https://github.com/Kademi/keditor#readme",
   "devDependencies": {
     "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.2.1",
+    "@babel/plugin-transform-modules-umd": "^7.2.0",
     "@babel/preset-env": "^7.2.0",
     "gulp": "^4.0.0",
     "gulp-babel": "^8.0.0",
@@ -27,6 +29,7 @@
     "gulp-cssmin": "^0.2.0",
     "gulp-header": "^2.0.7",
     "gulp-less": "^4.0.1",
+    "gulp-order": "^1.2.0",
     "gulp-plumber": "^1.2.1",
     "gulp-rename": "^1.4.0",
     "gulp-replace": "^1.0.0",
