+++ src/keditor/component/initDynamicContent.js
@@ -1,0 +1,43 @@
+import log from '../utils/log';
+
+export default function (dynamicElement) {
+    log('initDynamicContent', dynamicElement);
+    
+    let self = this;
+    let options = self.options;
+    let component = dynamicElement.closest('.keditor-component');
+    let contentArea = dynamicElement.closest('.keditor-content-area');
+    
+    dynamicElement.attr('id', self.generateId('dynamic-element'));
+    
+    if (typeof options.onBeforeDynamicContentLoad === 'function') {
+        options.onBeforeDynamicContentLoad.call(self, dynamicElement, component, contentArea);
+    }
+    
+    let dynamicHref = dynamicElement.attr('data-dynamic-href');
+    let data = self.getDataAttributes(component, ['data-type', 'data-dynamic-href'], false);
+    data = $.param(data);
+    log(`Dynamic href: ${dynamicHref}, Data: ${data}`);
+    
+    return $.ajax({
+        url: dynamicHref,
+        data: data,
+        type: 'GET',
+        dataType: 'HTML',
+        success: function (response, status, xhr) {
+            log('Dynamic content is loaded', dynamicElement, response, status, xhr);
+            dynamicElement.html(response);
+            
+            if (typeof options.onDynamicContentLoaded === 'function') {
+                options.onDynamicContentLoaded.call(self, dynamicElement, response, status, xhr, contentArea);
+            }
+        },
+        error: function (response, status, xhr) {
+            log('Error when loading dynamic content', dynamicElement, response, status, xhr);
+            
+            if (typeof options.onDynamicContentError === 'function') {
+                options.onDynamicContentError.call(self, dynamicElement, response, status, xhr, contentArea);
+            }
+        }
+    });
+};
