+++ src/keditor/constants/classNames.js
@@ -1,0 +1,8 @@
+const PREFIX_CLASS = 'keditor-';
+
+export default {
+    UI: `${PREFIX_CLASS}ui`,
+    HIDDEN_ELEMENT: `${PREFIX_CLASS}hidden-element`,
+    CLICKS_INITIALIZED: `${PREFIX_CLASS}clicks-initialized`,
+    CONTENT_AREAS_WRAPPER: `${PREFIX_CLASS}content-areas-wrapper`,
+};
