+++ src/keditor/constants/snippetType.js
@@ -1,7 +1,5 @@
-const SNIPPET_TYPE = {
+export default {
     ALL: 0,
     CONTAINER: 1,
     COMPONENT: 2
 };
-
-export default SNIPPET_TYPE;
