+++ src/keditor/getContent.js
@@ -1,0 +1,17 @@
+export default function (inArray) {
+    let self = this;
+    let result = [];
+    
+    self.contentAreasWrapper.find('.keditor-content-area-inner').each(function () {
+        let html = '';
+        $(this).children('.keditor-container').each(function () {
+            let container = $(this);
+            
+            html += self.getContainerContent(container);
+        });
+        
+        result.push(html);
+    });
+    
+    return inArray ? result : result.join('\n');
+};
