+++ src/keditor/utils/generateId.js
@@ -1,7 +1,5 @@
-const generateId = (type = '') => {
+export default (type = '') => {
     let timestamp = (new Date()).getTime();
     let random = Math.round(Math.random() * 9876543210);
     return `keditor-${type}-${timestamp}${random}`;
 };
-
-export default generateId;
