+++ src/keditor/utils/log.js
@@ -1,7 +1,7 @@
-const log = (...args) => {
-    if (console && typeof console.log === 'function') {
+import $ from 'jquery';
+
+export default (...args) => {
+    if (console && typeof console.log === 'function' && $.keditor.debug) {
         console.log.apply(console, ['[ KEditor ] ', ...args]);
     }
 };
-
-export default log;
