+++ src/test.html
@@ -89,17 +89,17 @@
         <script src="../examples/plugins/formBuilder-2.5.3/form-render.min.js"></script>
         <script src="../examples/plugins/bootstrap-3.3.6/js/bootstrap.min.js"></script>
         <script src="js/keditor.js"></script>
-        <script src="js/keditor-component-audio.js"></script>
-        <script src="js/keditor-component-googlemap.js"></script>
-        <script src="js/keditor-component-photo.js"></script>
-        <script src="js/keditor-component-text.js"></script>
-        <script src="js/keditor-component-video.js"></script>
-        <script src="js/keditor-component-vimeo.js"></script>
-        <script src="js/keditor-component-youtube.js"></script>
-        <script src="js/keditor-component-form.js"></script>
+        <script src="components/keditor-component-audio.js"></script>
+        <script src="components/keditor-component-googlemap.js"></script>
+        <script src="components/keditor-component-photo.js"></script>
+        <script src="components/keditor-component-text.js"></script>
+        <script src="components/keditor-component-video.js"></script>
+        <script src="components/keditor-component-vimeo.js"></script>
+        <script src="components/keditor-component-youtube.js"></script>
+        <script src="components/keditor-component-form.js"></script>
         <script type="text/javascript">
             
-            KEditor.debug = true;
+            $.keditor.debug = true;
             
             $(function () {
                 keditor1 = $('#keditor1').keditor({
